+++ db/updates/update_11.9-12.0.sql
@@ -1,5 +1,6 @@
-RENAME TABLE `lids` TO `accessories`;
-ALTER TABLE `accessories` CHANGE `style` `name` VARCHAR(255) NOT NULL; 
-ALTER TABLE `accessories` CHANGE `colour` `accessory` VARCHAR(255) NOT NULL; 
-ALTER TABLE `accessories` CHANGE `supplier_link` `supplier_link` VARCHAR(255) NOT NULL;
-ALTER TABLE `accessories` CHANGE `supplier` `supplier` VARCHAR(255) NOT NULL; 
+RENAME TABLE `lids` TO `inventory_accessories`;
+ALTER TABLE `inventory_accessories` CHANGE `style` `name` VARCHAR(255) NOT NULL; 
+ALTER TABLE `inventory_accessories` CHANGE `colour` `accessory` VARCHAR(255) NOT NULL; 
+ALTER TABLE `inventory_accessories` CHANGE `supplier_link` `supplier_link` VARCHAR(255) NOT NULL;
+ALTER TABLE `inventory_accessories` CHANGE `supplier` `supplier` VARCHAR(255) NOT NULL; 
+ALTER TABLE `inventory_accessories` ADD UNIQUE(`name`);