+++ pages/views/inventory/compounds.php
@@ -153,13 +153,13 @@ while($res = mysqli_fetch_array($q)){
               <li>
                 <div id="raw" data-size="<?=getMaximumFileUploadSizeRaw()?>">Maximum file size: <strong><?=getMaximumFileUploadSize()?></strong></div>
               </li>
-              <li>Any ingredient with the same ID will be replaced. Please make sure you have taken a backup before importing a JSON file.</li>
+              <li>Any compound with the same ID will be replaced. Please make sure you have taken a backup before importing a JSON file.</li>
             </ul>
           </div>
         </div>
       </div>
       <div class="modal-footer">
-        <button type="button" class="btn btn-secondary" data-bs-dismiss="modal" id="btnCloseBK">Cancel</button>
+        <button type="button" class="btn btn-secondary" data-bs-dismiss="modal" id="btnCloseBK">Close</button>
         <input type="submit" name="btnRestore" class="btn btn-primary" id="btnImportCompounds" value="Import">
       </div>
     </div>