+++ pages/views/inventory/editAccessory.php
@@ -5,7 +5,7 @@ require_once(__ROOT__.'/inc/sec.php');
 require_once(__ROOT__.'/inc/opendb.php');
 
 
-$accessory = mysqli_fetch_array(mysqli_query($conn, "SELECT * FROM accessories WHERE id = '".$_GET['id']."'")); 
+$accessory = mysqli_fetch_array(mysqli_query($conn, "SELECT * FROM inventory_accessories WHERE id = '".$_GET['id']."'")); 
 $doc = mysqli_fetch_array(mysqli_query($conn,"SELECT docData AS photo FROM documents WHERE ownerID = '".$accessory['id']."' AND type = '5'"));
 $sup = mysqli_query($conn, "SELECT id,name FROM ingSuppliers ORDER BY name ASC");
 while ($suppliers = mysqli_fetch_array($sup)){