+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Lids inventory dropped to accessories
 - Minor overall UI updates
 - Migrating backend scripts under a common backend api
+- Import json functions update
+- Added import for accessories
 
 ### Version 11.9
 - Added system logs access via the UI for docker/cloud installations - this comes disabled by default